refactor: reorganize source into shared and feature-based modules
