+++ PROJECT_STATUS.md
@@ -1,0 +1,7 @@
+# PROJECT_STATUS (Moved)
+
+프로젝트 현황 문서는 아래로 이동했습니다.
+
+- [`docs/status/current-status.md`](./docs/status/current-status.md)
+- 플랜 대비 비교: [`docs/plan/plan-vs-current.md`](./docs/plan/plan-vs-current.md)
+
