+++ bank/sakura_bank_ver_vue/src/features/transactions/stores/categories.ts
