+++ bank/sakura_bank_ver_vue/src/shared/lib/utils.ts
