+++ bank/sakura_bank_ver_vue/src/shared/ui/UiCard.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 const props = defineProps<{ class?: string }>();
 </script>
