+++ docs/plan/implementation-plan.md
@@ -1,0 +1,46 @@
+# MoneyBlosSum 구현 플랜 기준선 (Baseline)
+
+기준 문서: 팀 기획안 + KB 스켈레톤 가이드  
+기준일: 2026-04-07
+
+## 1. 필수 완료 범위 (MVP)
+
+1) 가계부 핵심
+- 수입/지출 등록 (날짜, 금액, 카테고리, 메모)
+- 거래 내역 조회 (필터/정렬)
+- 월별 수입/지출/순이익 요약
+- json-server 저장/수정/삭제
+
+2) 페이지
+- 메인(대시보드)
+- 캘린더(일별 내역 + 빠른 입력)
+- 차트 요약(월별/카테고리별)
+- 거래 등록/수정
+- 설정/프로필
+
+3) 기술 요구
+- Vue3 + Vite + Composition API
+- vue-router + axios + Pinia
+- 컴포넌트 기반 구조
+
+## 2. 권장 추가 범위 (확장)
+
+- 벚꽃 테마 인터랙션 (벚나무 성장/낙화)
+- 부적 시스템
+- AI 소식지
+- 위치 기반 소비 시각화
+- 영수증 OCR 입력
+
+## 3. 산출물
+
+- 소스 코드
+- 기획/화면 설계 문서
+- 발표 자료(PDF)
+
+## 4. 완료 정의 (DoD)
+
+- 핵심 CRUD가 json-server에서 정상 동작
+- 주요 페이지 라우팅 오류 없음
+- 월별 요약 값이 데이터와 일치
+- 최소 1회 빌드 성공 (`npm run build`)
+
