+++ docs/status/current-status.md
@@ -1,0 +1,20 @@
+# 현재 진행상황 체크
+
+기준일: 2026-04-07
+
+## 상태 요약
+- 앱 골격(라우터/레이아웃/스토어) 구축 완료
+- 거래 CRUD(json-server) 구현 완료
+- 일부 페이지는 목업 데이터 사용 중
+
+## 빠른 체크리스트
+- [x] 거래 등록/조회/수정/삭제
+- [x] 카테고리 조회 연동
+- [x] 프로필 저장 연동
+- [x] 빌드 성공
+- [ ] 로그인/권한 흐름
+- [ ] 전용 로딩 페이지
+- [ ] 캘린더 실데이터화
+- [ ] 모달 입력
+- [ ] 독립 차트 페이지 고도화
+
